refactor: extract distributed dispatch; rename opaque photon stages
__call__'s body was an isolated call-graph community: chunk sizing, cluster
lifecycle, map_blocks fan-out, and the progress bar -- pure execution
machinery with no physics edges. It now lives in dispatch.py as
map_showers_distributed(chunk_fn, ...); CphotAng.__call__ keeps only the
physics packing (its run() closure) and delegates. All dask/rich imports
leave cphotang.

The three "photon" stages whose names didn't distinguish them get honest
ones: _photon_yield -> _transmitted_yield (per-node transmitted yield
weights), _photon_sum -> _kernel_energy_integral (the energy-quadrature
kernel evaluation), _photon_density -> _density_at_detector.

cphotang.py: 1583 -> 958 lines across the six moves. Goldens bit-identical.
